voice/requirements.txt
elevenlabs==1.13.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0
httpx==0.27.0
